+++ CHANGELOG.md
@@ -10,11 +10,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 <!-- ### Added | Changed | Deprecated | Removed | Fixed | Security -->
 
+<!--------------------- Don't add new entries after this line --------------------->
+
+## [0.2.3] - 2023-12-15
+
 ### Added
 
 - `runner_pod_tpl` option for better control over the runner pod manifest [#2](https://github.com/mruoss/flame_k8s_backend/pull/2)
+- Basic integration test
 
-<!--------------------- Don't add new entries after this line --------------------->
+### Changed
+
+- Delete pod when shutting down the runner.
 
 ## [0.2.2] - 2023-12-14